adding WKCookie type to set cookies as is
setting all the cookies as a list of strings and parsing with soup_cookie_parse
doesn't always work as soup_cookie_parse adds a dot to the beginning of the domain so
saving cookies from the outside with getCookies and setting them back with
setCookies domain is modified changing the application of the cookie.

Passing as a WKCookie/WebCore::Cookie domain can be set as is.