[WebRTC] Support RTCIceServer.url along with RTCIceServer.urls
The current status of the spec says that "urls" should be used,
but previously the name of this paratemer was "url".
And it seems major services (like AppearIn) still use "url"

Chromium and Firefox support both names for this parameter.
See https://crbug.com/659133 and https://bugzil.la/1116766

This makes this test pass: http://iswebrtcready.appear.in/

Upstream-Status: pending