[GTK] Always allow usage of mediaDevices.enumerateDevices()
This implements the UIClient::checkUserMediaPermissionForOrigin() method,
which is called for checking for permissions when enumerating media devices.
For now access is always allowed, to avoid having to implementing a new kind
of permission request in the WebKitGTK+ API. Anyway, the documentation in MDN
states that navigator.mediaDevices.enumerateDevices() may be used without a
permission request and even acknowledges that this can be used by websites for
fingerprinting users. It would be nicer to ask the user for permission, but
it's not like this is doing anything against the spec...

* UIProcess/API/gtk/WebKitUIClient.cpp: