[EME] Attach the CDM instance to CDMClient in generateRequest
Currently the CDMInstance is attached to the CDMClient (HTMLMediaElement) when setMediaKeys is called.

But when we call setMediaKeys in javascript, we don't know if the MediaPlayer is created yet, thus when we attach
the CDMInstance to the CDMClient, we are not sure that the CDMInstance is propagated to the MediaPlayer and then to
MediaPlayerPrivate.

The fix constist in attaching the CDMInstance in generateRequest because we are sure at this moment that the MediaPlayer
and MediaPlayerPrivate is already created.